Refactor Localization class (#1308)
* refactor Localization classes

* update some comments

* use std ptr; bugfix in CMakeLists for C++14 std; update comments

* Remove time argument from Localization constructor (#1314)

* Removing time argument in Localization constructor

* Last occurence removed
