+++ l95/src/lorenz95/instantiateLocalizationFactory.h
@@ -10,13 +10,12 @@
 
 #include "lorenz95/L95Traits.h"
 #include "lorenz95/LocalizationMatrixL95.h"
-#include "oops/interface/Localization.h"
+#include "oops/interface/LocalizationBase.h"
 
 namespace lorenz95 {
 
 void instantiateLocalizationFactory() {
-  static oops::LocalizationMaker<L95Traits,
-         oops::Localization<L95Traits, LocalizationMatrixL95>> makerL95_("L95");
+  static oops::interface::LocalizationMaker<L95Traits, LocalizationMatrixL95> makerL95_("L95");
 }
 
 }  // namespace lorenz95
