+++ qg/model/instantiateQgLocalizationFactory.h
@@ -11,7 +11,7 @@
 #ifndef QG_MODEL_INSTANTIATEQGLOCALIZATIONFACTORY_H_
 #define QG_MODEL_INSTANTIATEQGLOCALIZATIONFACTORY_H_
 
-#include "oops/interface/Localization.h"
+#include "oops/interface/LocalizationBase.h"
 
 #include "oops/qg/LocalizationMatrixQG.h"
 #include "oops/qg/QgTraits.h"
@@ -19,8 +19,7 @@
 namespace qg {
 
 void instantiateQgLocalizationFactory() {
-  static oops::LocalizationMaker<qg::QgTraits,
-               oops::Localization<qg::QgTraits, LocalizationMatrixQG> > makerQG_("QG");
+  static oops::interface::LocalizationMaker<qg::QgTraits, LocalizationMatrixQG> makerQG_("QG");
 }
 
 }  // namespace qg
